+++ SpatialGDK/Source/SpatialGDK/Public/Interop/Connection/SpatialWorkerConnection.h
@@ -5,6 +5,8 @@
 // TODO(Alex): move include to .cpp
 #include "RealWorkerConnection.h"
 
+#include "SpatialWorkerConnectionCallbacks.h"
+
 #include "SpatialCommonTypes.h"
 
 #include "Interop/Connection/ConnectionConfig.h"
@@ -22,25 +24,6 @@ namespace SpatialGDK
 }
 class RealWorkerConnection;
 class FRunnableThread;
-
-UCLASS()
-class SPATIALGDK_API USpatialWorkerConnectionCallbacks : public UObject
-{
-	GENERATED_BODY()
-
-public:
-	DECLARE_MULTICAST_DELEGATE_OneParam(FOnEnqueueMessage, const SpatialGDK::FOutgoingMessage*);
-	FOnEnqueueMessage OnEnqueueMessage;
-
-	DECLARE_MULTICAST_DELEGATE_OneParam(FOnDequeueMessage, const SpatialGDK::FOutgoingMessage*);
-	FOnDequeueMessage OnDequeueMessage;
-
-	DECLARE_DELEGATE(OnConnectionToSpatialOSSucceededDelegate)
-	OnConnectionToSpatialOSSucceededDelegate OnConnectedCallback;
-
-	DECLARE_DELEGATE_TwoParams(OnConnectionToSpatialOSFailedDelegate, uint8_t, const FString&);
-	OnConnectionToSpatialOSFailedDelegate OnFailedToConnectCallback;
-};
 
 // TODO(Alex): use MoveTemp for function arguments
 // TODO(Alex): inline trivial functions
